cdl_utils: avoid specifying sc's on non-rt kernel
Add a flag to capdl_linker to indicate if we are building for an RT
kernel or not.

Create all threads as passive if this is a
non-rt build, to avoid sc's being inferred and placed into the spec.
This avoids a scenario where sc's are created in the spec for a non-rt
kernel and the capdl loader fails with an error.